Don't build same revision twice

+++ llvm/windows-build.sh
@@ -4,6 +4,16 @@ version=3.6.0
 
 pull . tools/clang
 
+git log -1 --format="%h" > .newbuild
+git --git-dir=./tools/clang/.git log -1 --format="%h" >> .newbuild
+cmp .newbuild .oldbuild &> /dev/null
+
+if [ $? = 0 ]; then
+    echo "No update, will sleep for 10 minutes and exit..."
+    sleep 10m
+    exit 0
+fi
+
 rm -rf dist
 mkdir dist
 cd dist
@@ -20,3 +30,5 @@ scp dist/LLVM-*.exe i10z.com:/havana/llvm/win32/LLVM-$version-$rev-win32.exe
 scp dist/build.log i10z.com:/havana/llvm/win32/build-$version-$rev.log
 ssh i10z.com ln -sf /havana/llvm/win32/LLVM-$version-$rev-win32.exe /havana/llvm/win32/latest.exe
 ssh i10z.com ln -sf /havana/llvm/win32/build-$version-$rev.log /havana/llvm/win32/latest.log
+
+mv .newbuild .oldbuild
